Renamed components directory.
