+++ ts/Extensions/DashboardPlugins/DataGridComponent.ts
@@ -22,7 +22,7 @@
 import type Options from '../../Core/Options';
 import type DataGrid from '../../DataGrid/DataGrid';
 
-import Component from '../../Dashboards/Component/Component.js';
+import Component from '../../Dashboards/Components/Component.js';
 import DataConverter from '../../Data/Converters/DataConverter.js';
 import DataStore from '../../Data/Stores/DataStore.js';
 import DataGridSyncHandlers from './DataGridSyncHandlers.js';
